Fixed cancel button behaviour

+++ src/main.py
@@ -177,10 +177,11 @@ class MainApp(qtw.QApplication):
     def cancel_patcher(self):
         self.patcher_thread.terminate()
 
-        if self.patcher.ffdec_interface._pid is not None:
-            os.system(f"taskkill /F /PID {self.patcher.ffdec_interface._pid}")
-            self.log.info(f"Killed FFDec with pid {self.patcher.ffdec_interface._pid}.")
-            self.patcher.ffdec_interface._pid = None
+        if self.patcher.ffdec_interface is not None:
+            if self.patcher.ffdec_interface._pid is not None:
+                os.system(f"taskkill /F /PID {self.patcher.ffdec_interface._pid}")
+                self.log.info(f"Killed FFDec with pid {self.patcher.ffdec_interface._pid}.")
+                self.patcher.ffdec_interface._pid = None
 
         if self.patcher.tmpdir is not None:
             if self.patcher.tmpdir.is_dir():
